Updates README with pipsi and pip install instructions.
pipsi is better for globally installed tools since it creates a sandbox
per python package you want to install. In this case, it installs the
zulipterminal package and all of its dependencies into one virtualenv,
then exposes the console script, `zulip-client`, for use even when that
virtualenv is not active, choosing to activate it only in the context of
that command.