run/core: Minimal fix for being initially unable to find server.
This minimally fixes #216 via the changes listed below.
This does not avoid an issue if other threaded methods fail due to a
lack of connection, particularly if they throw an exception, since
exceptions don't work well between threads.

This is not ideal, since we lose a thread option here, but it does
detect the lack of server connection early.

* core: Add ZulipServerConnectionFailure exception to core (later to model?)
* core: Remove @asynch from Controller.register_initial_desired_events
* run: If Controller() fails due to connection, catch this exception & exit