README: Improve pipenv notes.
We require python 3.4+, so we should be able to assume pip is present,
so we can just list the simple pip3 install of pipenv, rather than using
curl/get-pip/patching-PATH/etc.

Also clarify that `pipenv --three` initializes the venv, and subsequent
commands install into it.

Lastly, using `pip3 -e` inside of the venv is better for consistency
than `setup.py develop`. Thanks to @punchagan for testing this; this
only potentially causes an issue when old egg files are present.