TopButton: Use 'any' wrap for overflowing texts.
Don't space wrap the last word if it doesn't fit in the column.
Instead fill part of word in one row and rest in the next row.
This is done to keep computation of showing unread count at
the right position simple. Otherwise, we will have to take
wrapping of words into account which is currently not included
in the code.