travis/README: State we support running on PyPy and test on PyPy 3.5.
Running of zulip-term under PyPy appears to work fine, tested manually
in a docker environment, so add classifiers to support PyPy (and
CPython).

One caveat for testing is that some libraries required for development
do not run with PyPy; specifically, this limits running mypy with PyPy.
Since Travis currently uses pipenv to install packages, independently
of the python version, this commit removes mypy from the Pipfile and
adds an additional instruction to the README.md to install it
separately if using pipenv (via requirements.txt, to avoid duplication).
Then we can add testing against PyPy to Travis (just version 3.5 for
now).