boxes: Avoid newline characters at start of blockquotes.
The incoming message (soup) for blockquotes contains '\n'
characters at various places. Some of those which occur later
on are useful (act as separator) while the first few used to cause
a large gap at the start of the message, varying with the level of
blockquoting.

This commit removes the latter newline characters to conserve space at
the start of messages with quoting.