boxes: Avoid newline characters at start of blockquotes.
The incoming message (soup) for blockquotes contains '\n'
characters at various places. Some of those which occur later
on are useful (act as seperator) while the first few used to cause
a large gap at the starting of the message.

Now, we only allow the later occuring newline characters to
conserve space;