views: Build TopicView by building TopicButtons of a stream.
We build the TopicView for a particular stream on activating its
StreamButton. The topics stored in the index allow building a
TopicButton for each of them, to be passed to the TopicView __init__.

Finally, the TopicView is wrapped with a LineBox; the top-line of
topics_view is made bold as per the UI change introduced in 16a9aea.

Tests added.