ui: Restrict max width of center column for easier reading.
Adds a config option layout=space-between to make the center
column fix width and add space between the three columns on
wider screens. On smaller screens, the original behavior is
preserved.

This mimics the webapp's idea of having the line-length be
small so that the text is easy to read even on larger screens.