model/views: Update topic order in open TopicView.
When TopicView of a particular topic is toggled open, we
show topics belonging to a stream in the LeftColumnView.
If a new message arrives now, we need to update the topic
order based on the new message unconditionally.

If no previous topics with the same topic names are found
we create a new TopicButton for it.

The logic here is similar to Model.update_topic_index.

Tests added for update_topics_list.