buttons: Mark 'muted' topics with 'M' when displaying topic view.
When we open TopicView for a particular stream, muted topics are now
shown with an 'M', as for muted streams.

NOTE: As per the TODO, the muted topic status does not currently update.

Tests amended for test_init_calls_top_button and
test_update_topics_list.

Test added for calling mark_muted when the topic encountered is
muted.