model: Re-order topic_index based on the topic of most recent message.
During an append_message event for streams, we need to update
the topic order stored in the index. This has to take place
unconditionally regardless of whether we have fetched most recent
message, or UI has loaded.

We create a new entry and place it in index 0 of index_list if this is a
new topic.

Test added for update_topic_index.