bugfix: Ensure 'self.update' property is set after messages are fetched.
On startup, the event handler for tracking incoming messages would
be registered with 'append message' as callback. However, the 'self.update'
variable would not be set to true yet. This is due to 'get_messages' not
setting the parameter within the conditional when 'query_range'
and length of response are equal.

Tests added to identify this.