bugfix: boxes: Avoid indenting single blockquotes within same message.
This corrects one situation arising from the quoting enhancement in
e7421af, where we treated a single message with multiple 'single'
blockquotes as being quoted multiple times (level 2,3..so on).

This commit avoids this by resetting the pad_count to 0, thereby
maintaining it at 1 (due to inc at the end) in each iteration of the
'single' blockquote message.