core: Add support for opening any link in the web browser.
This adds open_in_browser() to add support for opening any link in the
web browser. This can later be used to either view any message or open
an external link in the browser.

The open_in_browser() accepts a URL and uses webbrowser module to open
that in browser. If no runnable browser is found in the system running
ZT, then an appropriate error message will be displayed in the footer.

Test added.

Co-authored-by: Puneeth Chaganti <punchagan@muse-amuse.in>
Co-authored-by: Preet Mishra <ipreetmishra@gmail.com>
Co-authored-by: Zeeshan <equbalzeeshan@gmail.com>