bugfix: Update unread count for messages which cannot be indexed.
For narrows, in which not all the messages have been fetched yet,
when a new message arrives, it cannot be appended to the view (index).
However, even its unread count does not increase. To avoid this,
we maintain an un-indexed message collection.

When, a new message arrives, we add its id to the un-indexed collection
and increment its count.

When we narrow to a stream, or load_new_messages, narrow_to_user, etc,
we pop the fetched message which are un-indexed as they
are now indexed.

The indexed and un-indexed collections are mutually exclusive to each other.

Tests amended.

Fixes: #395