update_count: Inc/dec 'all_msg' count when a stream is muted/unmuted.
When a stream is muted, we remove the unread count of the stream
from the 'all msg' button. When it gets unmuted, if it still contains
unread messages, we restore it back and increase the 'all msg' count
by the same amount.

Earlier tests amended by adding new parameters to test
'all_msg' count changes.