utils: Avoid displaying messages from unsubscribed streams.
We build the stream dict with data from streams the
user is subscribed to. When a message belonging to an
unsubscribed stream is fetched (When we unsubscribe from a
stream in web-app and open ZT), it causes a crash due to
the 'stream_id' of the message not exisiting in stream dict.

To avoid this, we do not display 'unsubscribed' messages
during creation of message box list.

Tests added.