refactor: core/views: Extract show_pop_up method from show_help.
The help menu is now shown using the new method, by providing the widget
and popup title text.

As part of extracting the show_pop_up method, the attribute of the
widget used for sizing has been adjusted from number_of_actions to
height, which is more meaningful for a general pop-up.