helper: Ensure set_count() updates the model for all_msg & all_pms.
Unread counts in the model were previously only updated for specific
private messages (pms) and topics.

This commit ensures the unread counts for all_msg and all_pms are now
also updated in both the model as well as the UI.

NOTE: We do not yet address updating the remaining entry in the unread
counts, ie. streams.