helper: Add function for canonicalizing colors.
The zulip server used to store some colors in the format #xxx and
others in the format #xxxxxx. The terminal client is currently not
handling the #xxx format correctly in all cases.

This commit adds a function for taking **either** server format and
converting it into the correct client format.

Tests added for canonicalize_color.