model/views: Set focus after updating button order in TopicView.
When a new message comes in and the topic_view is full, the
most recent topic occupies the top position. To ensure the
focus is set on the topic, we 'scroll-up' or set valign to
make the top-most topic button always visible.

Further, if this is a message whose author is you, we set the
focus to point at the most recent topic button of the message
you just sent.

Tests amended.