helper: Ensure set_count() updates the model for muted & unread_topics.
This commit ensures the unread_counts for muted msgs and unread topics
are now also updated in the model.

Updating model for reading muted events is useful when we read some
messages in muted stream and decide to mark the stream non-muted. We
are now able to restore the correct unread_counts.