model/helper: Fix conditional for adding/fetching messages from index.
Helper aded to check if a  narrow is a result of searching. This can be
determined by checking if the narrow containts ['search', text] at any
position.

We make use of the helper to now as the conditional to add and fetch
searches messages in the index.

Tests added for model.is_search_narrow().