model: Register realm_user_groups during initial data fetch.
We register realm_user_groups event type during initial data fetch
to get information regarding user_groups in realm.

We pass the information to a private method to extract
and store the information in model to be used later.

We return the sorted list of group_names and store it in model
as well so that it can be made use of in the typeahead search.

Tests amended for conftest and TestModel.

Tests added for _group_info_from_realm_user_groups.