helper: set_count: Adjust count of group PMs.
If a message is a group PM (huddle), adjust its unread count stored in
unread_huddles.

NOTE: This also is a bugfix to an earlier split-out bugfix commit
(49b546b), since after that commit the key/unreads variables were
undefined in the case of huddles, due to the absence of an 'else'
branch covering that case.