model: Register handler for muted_topic events for (un)muting.
We want to be notified about changes in subscriptions like
muting/unmuting of topics in a stream, hence we register the
event_type 'muted_topics' to the event queue on startup, with a
handler method which responds to these events appropriately.

We store mappings between topic_name  and topic_buttons so
tha retrieval of buttons is easier when a change in topic muting
event occurs.