tests: model: Refactor notify test to use fixture message details.
Previously the sender_id and user_id were adjusted in the test, to
ensure notifications aren't sent when the user is the sender, but are
otherwise. This commit refactors the test to use the ids in the
message_fixture itself but achieve the same result. This isn't ideal, as
it is more coupled to the fixture, but without this then more detailed
testing requires other message_fixture elements to be updated too.

In future it might be useful to have the message_fixture be consistent
with the other user ids now present in the other fixtures.