requirements: Upgrade zulip from 0.5.9 to 0.6.1.
The new zulip library version allows catching of ZulipError in an
additional case, removing some cases where a requests exception was
thrown directly into zulip-terminal.

This avoids an uncaught exception, in cases such as being connected to a
network but not routed, such as if not signed in to a wifi router.