refactor: model: Use str for ThreadPool function return values.
This allows for more information to be returned beyond the success/fail
from a bool, particularly useful in the case of failure.

Return values are changed as follows:
* True -> ''
* False -> 'Some error response'

Note that this flips the truthiness of the return type, leading to
addition/removal of negation to retain previous behavior.

Tests and typing amended, including addition of a missing 'type:' in a
variable type hint.