model: Report error details from failed initial API calls.
Given failed calls now provide a textual reason, this information can be
passed to returned exceptions, and ultimately to the user.

This change applies to the initial connection only.

Since ZulipError error texts are now returned, as well as responses from
successful network calls which fail at the server, a test case is added
for each.