boxes/themes: Add inline emoji styling similar to reaction styling.
Emojis in messages are styled to be similar to the styling of reactions.
To differentiate between reactions and inline emojis, reactions are made
bold. A 'reaction' style was added to the theme, this allows better
customization of emojis and reactions independent of each other.

Tests amended (test_soup2markup for emoji styles).

Fixes: #412