tools: Sort imports with isort & enforce linting via Travis.
This reformats import statements throughout the code as per isort, and
ensures the formatting is retained by adding the new linting script to
Travis.

Travis runs isort on the lowest python version, since the 'atomic'
option should ensure valid syntax, which is assumed to be compatible
with other supported versions.