tools: Filter warnings from pytest/pep8 due to external libraries.
This cleans the output as seen by developers, to make it more obvious
that this is not an issue with ZT code, but rather external
libraries/tools.

Outstanding warnings hidden by this commit:
* pytest-pep8 doesn't set up some kind of 'hook' to register a marker
* imp module is deprecated in favor of importlib (used by distutils)
* collections.abc should be used for ABCs (used by bs4)

This commit works around these issues in setup.cfg by adding
filterwarnings entries for external deprecated usage, and explicitly
mentioning the pep8 mark, accompanied by notes to ensure these reasons
are not forgotten.