tools: Use pycodestyle in place of outdated pytest-pep8.
This is intended to be a step towards updated PEP8 linting, as now
provided by pycodestyle.

Other than moving to a more maintained and current linter (pycodestyle
vs pep8), using a separate linter makes it more obvious which aspects
are failing (tests vs pep8 linting) and reduces the need to run it in
every test run - including each version of python on CI. It also allows
the linter to be run consistently in editors more frequently (if
desired), as it runs fairly rapidly.

Locally this broadly just splits the time taken to run, with pycodestyle
taking ~3s.

* Update dependencies (pytest-pep8 -> pycodestyle)
* Add pycodestyle configuration to setup.cfg
* Remove pytest-pep8 configuration from setup.cfg
* Minimal changes to two files to show consistency
* Explicitly add PEP8 checks to only one CI task in travis
* Update README for different lint commands

Currently this explicitly excludes E252, for spacing around typed
function parameters, which is a newer style. This also excludes a newer
E117 check, which appears flakey.

All other excludes are as per pycodestyle defaults.