travis: Split linting into separate sections.
This makes it clearer to developers which CI tests/linting are failing.

Most are run on Python 3.7, since that is installed by default and so
does not require an extra python install step - so is faster. Linting
imports is not, as the tool should run on the lowest version supported
to ensure compatibility.

This was not done previously since the overhead from pipenv setup
(including locking) made setting up each environment far slower than
each test.