travis/setup: Clarify install requirements for CI testing & linting.
The different types of testing requirements are set in setup.py, with
two additional extras_require enabling smaller requirements for pure
linting and testing.

While the packages are cached, this also reduces the need to install the
remaining dev_helper packages on CI, and also only the linting packages
if linting is enabled in a job.