model: Handle read events for indexed messages.
This commit primarily adds handling of updates to the status flags on
messages for 'read' events, in addition to the previously-present
'starred' events, in update_message_flag_status.

We previously sent a message to the server indicating a particular
message had been read locally, also marking it read and updating
unread-counts locally. Now that we handle the 'read'-flag events, we can
also handle remotely-read messages, from other clients.

For now, the count updating is moved from the server-message method to
update_message_flag_status.

NOTE: One current limitation to updating remotely-read messages is that
we only handle messages we have full content of in the index.

Star-flag update tests amended.

New tests added for updating the read-flag.