boxes: Merge date/time and optional prev-year fields to improve styling.
While not visible in the default theme, in others the two parts of the
date are separated by a gap, due to how the optional year feature was
added as an extra column. By merging the columns and explicitly adding
spaces this now displays more cleanly and appears as one block of text.

Tests amended.