README/helper: Support notification sound for OSX.
The `osascript` command takes in the notification sound as an
extra parameter and without it doesn't have a default sound.
We set an env variable here to tell zt to play a
sound when it receives a notification.

README has been updated to describe the changes needed to
set the env variable.