tests: model: Combine _handle_reaction_event_add/remove tests.
This parametrizes over add/remove 'op' tags in the reaction events, and
tests the number of expected reactions after the event for each.

The existing other assertions are maintained for each of add/remove,
except removing the potentially complex check for the specific emoji
being present or not, and replacing with a check for the number of emoji
present after the event.