linting-fixup: boxes: Remove unused typing import.
The change to show 'my reactions' in 3b7e55097 switched from
typing.Counter to defaultdict, but didn't remove the import. The change
to flake8 in 29b3284fb enforced removal of unnecessary imports, but
wasn't rebased before merging, so briefly broke the flake8 task in CI.
This commit fixes this issue.