boxes: Render ordered lists fully with '  1. ' prefix.
Numbering starts with the first number specified in the list, as per the
Zulip revised scheme:
https://github.com/zulip/zulip/blob/master/docs/subsystems/markdown.md#lists

soup2markup now has a kw-args parameter 'state', in which a 'list_index'
entry is set by the ordered-lists code to have an index, incremented
through each list item, starting at the value specified in the 'start'
attribute of the html ordered list (ol).

Tests added.