model: Pass exception in event handlers to main thread.
Before this commit, exceptions in event handlers would result in a
traceback overlaid on the UI (if loaded); the application could appear
to continue as normal, but the event loop would fail - and identifying
tracebacks was challenging, as the UI might draw over the details.

This commit ensures that exceptions from event handlers are sent to the
main thread in the controller, using the infrastructure added in the
prior commit, where they are re-raised, causing the application to exit
and the exception to be logged.