requirements/linting: Add flake8-continuation plugin & enforce linting.
This plugin enforces one rule only: to avoid continuation characters;
consequently there is no configuration necessary in setup.cfg to add
this rule or exclude any parts of the plugin.

The motivation for this is given in part of PEP8:
https://www.python.org/dev/peps/pep-0008/#maximum-line-length

While there are special cases, at the time of commit this includes:

"The preferred way of wrapping long lines is by using Python's implied
line continuation inside parentheses, brackets and braces. Long lines
can be broken over multiple lines by wrapping expressions in
parentheses. These should be used in preference to using a backslash for
line continuation."