boxes: Indicate active search results for streams/topics/users.
Each of the streams/topics/users views use the PanelSearchBox to handle
search queries, but previously have not made it particularly clear when
a search is currently limiting display of entries (eg. streams).

This commit shows/hides a caption indicating whether a search is in
progress, in a similar way to how the
MessageView/MiddleColumnView/SearchBox does so, ie. with a styled
'Search Results' text.

Tests amended.

Partial fix towards #552.