boxes/themes: Set recipient bars to be thick colored lines.
Following recipient names, eg. stream/topic or PM recipients, we
previously had a full-height bar covering the width of the message-list.
This shrinks the height of that bar to be a thick line, also now colored
if associated with a stream.

For non-stream narrows which use the 'general_narrow' style, this adds a
new 'general_bar' style for styling the lines adjacent to this text.

Tests amended.