boxes: Pack recipient headers to ensure they are displayed.
This commit allows recipient headers to wrap onto multiple lines, rather
than being a fixed width, which previously caused them to not be
rendered in narrow windows.

This at least shows the recipient header if they are long, however the
styling could be improved. A custom urwid widget may be necessary to
improve further.

Fixes #671.