tests: model: Add failing test for unread count bug.
While handling the message event, message indexing and unread count updates
should happen even if user in the same narrow as that of the received
message. `model.found_newest` is True when the received message belongs
to the current narrow.

Currently, a message is added to index and unread counts are updated only
when `model.found_newest` is True. This leads to
bugs like negative unread counts and missing message information in the
index. This commit adds a failing test for fixing this bug.