boxes: Use styled space in recipient header between recipients and line.
This improves consistency between how private and stream message
recipient headers are shown.

For private messages, this avoids the space being spuriously highlighted
(selected) when the message underneath is selected, as a separate space
in the columns was not previously styled.

For stream messages, this ensures that the stream/topic text is
separated from the line/bar by a space, if it wraps onto a second line;
previously the space was part of the stream/topic text.

Tests amended.