views: Show author name and author prefix in EditHistoryView.
This extends _make_edit_block() to show the author name and adds
_get_author_prefix() to increase its verbosity by indicating the changes
that were made in a particular version (e.g. Content, Topic, Content &
Topic).

The implementation has a workaround to spot a false alarm for content
or topic change (A NOTE comment has been added) and an author prefix,
'Edited but no changes made', for edits that do not have any 'real'
changes.

Test amended and added.