refactor: boxes: Standardize autocomplete_* signatures for consistency.
This standardizes the autocomplete_streams and autocomplete_emojis
methods to accept prefix_string as an argument like the other
autocomplete method.  However, unlike the other autocomplete method,
this doesn't use prefix_string in the method bodies (deliberately).

The intention is to lay the groundwork for generalizing other methods
that use these autocomplete_* methods.