conftest/core/views: Show server version and feature level in About.
The server version and the feature level are shown under a 'Server'
category. Note that the feature level is only shown when it is present.

For testing, this also introduces zulip_version fixture to test
different components based on the server version and the feature level.

Test amended and added.

Fixes #616.