ui/boxes: Display typeahead suggestions in footer.
Display the first 10 suggestions in the footer when user presses
the autocomplete key. The current suggestion is highlighted and
' [no matches found] ' is displayed when there are no matches.
If there are more matches than currently displayed, '[more]' is added to
the end of the footer text.

generate_typeaheads has been modified to return both the names of the
elements and the styled typeaheads. The name is used to display the
footer text.

Tests added.