boxes: Add typeahead mode attribute for WriteBox.
Currently, there is no way to figure out when a user is done with
autocomplete, making it difficult to reset the footer. A duration based
mechanism for resetting the footer is not ideal because multiple presses
of AUTOCOMPLETE_KEY will lead to multiple changes in help text.

This commit adds a boolean variable (is_in_typeahead_mode) to determine
when the user is done using typeahead. If the user is not using typeahead,
default footer text is rendered.

is_in_typeahead_mode is set to True on when user presses any of the
autocomplete keys, and set to False when the user pressed any other key.