README/version: Indicate supported (and minimum) server versions.
This adds SUPPORTED_SERVER_VERSIONS and MINIMUM_SUPPORTED_SERVER_VERSION
in version.py. Additionally, this adds a note in the README to specify
the minimum supported server version.

The intent is to use these versions for testing and potentially provide
feedback to the user if the server version is below the
MINIMUM_SUPPORTED_SERVER_VERSION.