boxes: Enter on a empty string cancels the search
An empty search string shows an empty search box and
jumps to eg. the full selection of streams,
leaving the search box blank. This caused confusion because
it was an empty search.

This fix causes an empty string search for streams/users/messages,
with enter, to now act the same as 'esc';ie. it
cancels the search and returns to the full list of streams/users
/messages.
A test for this behavior was also added.

Fixes: #551.