model: Make content optional in update message request.
The message content for updating stream message should not be passed to
the request when it is not supposed to be edited i.e. when the time
limit for editing it has expired. If it is, the server will return a 400
bad request. This behaviour has not been updated in the REST API docs as
of this commit, but can be confirmed at
https://github.com/zulip/zulip/blob/7229088b6cd331ed8370fb11625346f1a5614c2b/zerver/views/messages.py#L1588

This commit makes content an optional parameter in
model.update_stream_message.

This change is important when implementing topic-edit-only for messages.

The API supports `topic` instead of `subject` in the arguments since
server version 2.0 - this has been updated accordingly for the
update_stream_message endpoint. Also, renamed the msg_id argument to
message_id to make tests simpler.

Tests amended.