boxes: Allow editing of topics after message editing time limit expires.
This commits allows users to edit topics of their own message even if
the time limit for editing the message body has expired. This also
includes a UI change of disabling focus of write_box.msg_write_box when
such a situation arises. Note that this is allowed only when realm
editing for messages is enabled.

* Modify CYCLE_BOX_FOCUS keypress for WriteBox so that focus is not set to
  msg_write_box.
* Store an attribute in WriteBox to ascertain disabled/enabled
  state.
* Amend tests.