views: Add EmojiPickerView to add reactions to messages.
A new pop-up EmojiPickerView will be triggered on pressing `:` key on a
message. This view will be used to search and select the emoji that a
user wants to add to a message.

This commit includes a minimal view that only includes the 'thumbs-up'
emoji. Buttons for this emoji, and a search bar for more emojis will be
added in subsequent commits.

Tests added.