boxes/views/core: Add EmojiPickerView to add emojis to message.
A new pop-up EmojiPickerView will be triggered on pressing ':' key on a
message. This view will be used to search and select the emoji that a
user wants to toggle (react/unreact with). The selection will be applied
only after exiting the popup.

This currently only provides a UI for adjusting the thumbs-up/+1 emoji,
as currently achievable via `+`, but will be extended in future.

A minimal search-box implemented on top of PanelSearchBox is added,
which can be triggered on pressing 'p' key from Emoji picker view. The
search logic will be added in subsequent commits, however currently only
one emoji is in the list.

Tests added.

Co-authored-by: Zeeshan <equbalzeeshan@gmail.com>