model: Generalize react_to_message() to support all reactions.
Presently, the emoji code and type for 'thumbs_up' is hard-coded in
the react_to_message method. This commit uses the stored emoji data (in
unicode_emojis.py) to look-up the emoji code and type via
model.active_emoji_data. This change makes it easier to use
model.react_to_message, in the future, for emojis other than the
currently supported 'thumbs_up'.

As per the API docs, the "user" object within reactions field in message
response is deprecated and will be removed in the future. Instead of this
a new "user_id" field was created within reactions in server v3 (ZFL 2).
We add support for this field along with the "user" object so that the code
doesn't break in the future.

Tests added and amended.

Co-authored-by: Zeeshan <equbalzeeshan@gmail.com>