model: Use stored emojis to fetch emoji data for reactions.
Presently, the emoji code and type for 'thumbs-up' is hard-coded
in the react_to_message method. This commit uses the stored emoji
data(in emoji_names.py) to look-up the emoji code.

This change makes it easier to use model.react_to_message, in the
future, for emojis other than the currently supported 'thumbs-up'.