bugfix: views: Do not use an empty string to restrict user list update.
Previously, the method, update_user_list, relied on empty string ("") to
not update the user list (it assumed that urwid doesn't send signals for
an empty string) which caused it to not process and update empty string
user search.

This uses None instead of an empty string to set the restriction which
makes sure that empty search updates work properly.