model: refactor: Use repr() for finding model.narrow string.
repr() is safer to use for generating string representations since it
guarantees that it will yield an object with the same value when passed
to eval(). It is designed to be unambiguous, while str() is designed for
readability. str() leads to problems while using quotes, datetime
objects, etc.

Here, the string representation of model.narrow is used for storing the
pointer in the index. There is no visible change in the string
representation, but this refactor ensure future bugs should not arise due
to this.