core: Do not reset found_newest when changing narrow.
If found_newest is set to False on changing narrow, then new
messages will never be appending to that narrow. This is prevented by
the check in Model._handle_message_event(), whose purpose is to prevent
the appending of messages if the latest one isn't loaded yet for
pagination.  The only way that found_newest gets reset is in
Model.get_messages() where its value is returned by the server.

To reproduce this (now fixed) bug,
* Open stream A.
* Open stream B.
* Open stream A.
* Receive messages in stream A - which now won't be displayed.