bugfix: boxes/views: Ignore mouse click if compose box is open.
Clicking on messages when the compose box is open causes ZT to crash,
and having it open when clicking on user names causes the current message
to be lost. This commit ignores the mouse click to prevent this
from happening, though better handling can be done when the drafts
feature is implemented.

Also, stored controller as an attribute for UsersView. Tests amended for
this.

Test added for mouse click event.

Fixes #652.