bugfix: ui: Use a queue to manage footer reset durations.
Currently, if multiple calls are made to set the footer along with a
duration to reset it (for example, when tries to edit an uneditable
message too often by pressing the key too many times) then it leads to
repeated footer updates because all those calls lead to resetting the
footer. The goal of this commit is to change that, and reset the footer
only once, by ignoring other updates.

A queue is maintained containing number of sleep threads currently
active, and if there are none left then the footer is reset. This
ensures that only the last call to set_footer_text() resets the footer,
thus avoiding multiple resets.

Also, note that it is not strictly necessary to store the duration in
the queue, we could any number we like. However it is kept this way in
case we would like to do more with this in the future, say, spacing out
tight reset calls with only a small duration between them.

Test added.

Fixes #647.