boxes/model: Check if modified recipients are valid.
Users can manually alter the recipients/stream name, but currently,
there is no check in place to ensure validity of these modifications.
This commit checks the recipients everytime the user presses TAB on the
boxes (stream name and recipients for private messages), and displays an
error message in the footer if a recipient is invalid.