boxes: Delegate link text for raw links in a better way.
Currently, we use the last segment of a link for this purpose. It works
well for links that have an image or a file. For instance,
https://foo.com/image.png gets rendered as image.png, but it does have a
few caveats.

This shows the ending for a filename, the domain name for a general link
and the relative URL for an internal link.

Test amended.

Fixes #742.