model/views: Fetch topics on demand to prevent heavy load on /topics.
This shelves the concurrent fetch_all_topics() routine to prevent heavy
load on the /topics endpoint when the terminal app is started.

Now, topics are only fetched when they required for the topics_view()
(for the topic toggle) or for the _update_topic_index() via
topics_in_stream().

Tests amended (with a few spelling corrections) and added for
topics_in_stream().

Fixes #493.