boxes/model: Maintain stream_id in WriteBox for stream box view.
The stream id is useful for referring to a specific stream in a way that
is independent of its name, which can vary over time.

The stream_id needs to be synchronized in case a user manually modifies
the stream name in the recipient box.

This will be important for implementing message drafts and
autocompleting topics.

A new helper method is added to Model to fetch stream_id using its name.