refactor: model/boxes: Store emoji data in model.
Save the data from unicode_emojis.py as a model attribute,
instead of loading them directly from the file.

This will make it easy, in the future, to use emoji data when it is
fetched from different sources, for example - custom emojis that will
be fetched using a server endpoint.

Tests amended.