boxes: Disallow entering of Cc/Zs characters at start of PanelSearchBox.
The disallowed characters are:
Cc - Control characters
Zs - Space separators (noted as Zeparators in the code given the 'Zs')

Special casing is required to avoid how urwid treats eg. backspace not
as U+0008, but as a 'backspace' string.

After the first character, this just delegates to the default
implementation.

Tests added.