Wrap creation of zuliprc in try to catch OSError from open.
Previously OSErrors where caught in the outer try in the while loop in
parse_zuliprc. This resulted in a misleading error message about invalid
credentials. This patch catches OSErrors from open directly at the source,
and exits directly since there is no point to keep asking the user for
credentials. In particular, when running the default docker configuration,
open throws PermissionError since the zulip user in the container does not
have permission to create files in the mapped host folder.
Fixes #797.