buttons: Style text/marker for muted streams & topics distinctly.
This "dims" the stream/topic text and the muted marker for muted streams
and topics, which were previously styled just like regular streams and
unread counts.

Also includes minor bugfix, using just unread_count text as-is, as it
was changed previously to be a tuple.

Updated slightly by neiljp to store and use original text color, since
update_count is called in some cases without the text_color parameter
and the original color should then be used, not None.