keys/boxes/buttons/README: Add 'a' key for all messages.
This adds a new token 'ALL_MESSAGES' which maps to the 'a' key, as well
as the existing 'esc' key.

This replaces the 'GO_BACK' token ('esc' key) currently being used for
displaying all messages.

Now both 'GO_BACK' and 'ALL_MESSAGES' work, but only 'a' is shown in the
top-left menu.

Fixes #804