refactor: keys/model/boxes/buttons: Introduce primary_key_for_command().
keys_for_command currently returns a List and pop is used to get one key
to be displayed in the UI.

A new method primary_key_for_command is introduced, that always returns
the first element from the List, which is now intended to be the most
important (ie. primary) key.

pop is replaced everywhere with this new method.

tests are also added to assert this behaviour.